data to room

+++ app/src/main/java/fi/carterm/clearskiesweather/models/WeatherData.kt
@@ -7,8 +7,8 @@ import androidx.room.PrimaryKey
 data class WeatherData(
     @PrimaryKey(autoGenerate = true)
     val uid: Long,
-    val temp: Double,
-    val humidity: Double,
-    val pressure: Double,
-    val wind: Double) {
+    val temp: Float,
+    val humidity: Float,
+    val pressure: Float,
+    val wind: Float) {
 }
